requirements.txt
python-telegram-bot==21.7
openai==1.56.2
tiktoken==0.8.0
python-dotenv==1.0.1
google-genai
groq>=0.33.0
